refactor(main): extract pure helpers into per-feature modules (step 1)
Step 1 of #127. Move pure helpers out of the 2898-line src/main.rs
into focused modules and add a regression net of unit tests for
parse_args before later steps move command/processor code.

  src/cli/options.rs    Options, OutputFormat, StoredTagMode, AacAlbumInfo
  src/cli/parse_args.rs parse_args, expand_files_recursive (+ 27 unit tests)
  src/cli/usage.rs      print_usage, print_version, VERSION
  src/json_output.rs    JsonOutput, JsonFileResult, JsonAlbumResult, JsonSummary
  src/progress.rs       progress bar helpers + thresholds

No public library API change. CLI behavior unchanged.
