+++ src/main.rs
@@ -3,6 +3,10 @@
 //!
 //! Command-line interface compatible with the original mp3gain.
 
+mod cli;
+mod json_output;
+mod progress;
+
 use anyhow::Result;
 use colored::*;
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
@@ -12,18 +16,23 @@ use mp3rgain::mp4meta;
 use mp3rgain::replaygain::{self, AudioFileType, ReplayGainResult, REPLAYGAIN_REFERENCE_DB};
 use mp3rgain::{
     analyze, db_to_steps, delete_ape_tag, find_max_amplitude, read_ape_tag_from_file, steps_to_db,
-    undo_gain, Channel, GainOptions, GAIN_STEP_DB, TAG_MP3GAIN_MINMAX, TAG_MP3GAIN_UNDO,
+    undo_gain, Channel, GainOptions, TAG_MP3GAIN_MINMAX, TAG_MP3GAIN_UNDO,
     TAG_REPLAYGAIN_ALBUM_GAIN, TAG_REPLAYGAIN_ALBUM_PEAK, TAG_REPLAYGAIN_TRACK_GAIN,
     TAG_REPLAYGAIN_TRACK_PEAK,
 };
-use serde::Serialize;
 use std::env;
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::time::SystemTime;
 
-const VERSION: &str = env!("CARGO_PKG_VERSION");
-const PROGRESS_THRESHOLD: usize = 5;
+use cli::options::{AacAlbumInfo, Options, OutputFormat, StoredTagMode};
+use cli::parse_args::{expand_files_recursive, parse_args};
+use cli::usage::print_usage;
+use json_output::{JsonAlbumResult, JsonFileResult, JsonOutput, JsonSummary};
+use progress::{
+    create_analysis_progress_bar, create_progress_bar, finish_analysis_progress, progress_finish,
+    progress_inc, progress_set_message, update_analysis_progress, PROGRESS_THRESHOLD,
+};
 
 /// Extract filename from path, returning "unknown" if extraction fails
 fn get_filename(path: &Path) -> &str {
@@ -31,153 +40,6 @@ fn get_filename(path: &Path) -> &str {
         .and_then(|n| n.to_str())
         .unwrap_or("unknown")
 }
-
-// =============================================================================
-// Options
-// =============================================================================
-
-#[derive(Default, Clone, Copy, PartialEq)]
-enum OutputFormat {
-    #[default]
-    Text,
-    Json,
-    Tsv, // Tab-separated values (database-friendly)
-}
-
-#[derive(Default, Clone, Copy, PartialEq)]
-enum StoredTagMode {
-    #[default]
-    None, // Default behavior
-    Check,    // -s c: Check/show stored tag info
-    Delete,   // -s d: Delete stored tag info
-    Skip,     // -s s: Skip (ignore) stored tag info
-    Recalc,   // -s r: Force recalculation
-    UseApev2, // -s a: Use APEv2 tags (default)
-}
-
-/// Album gain info for AAC files
-struct AacAlbumInfo {
-    album_gain_db: f64,
-    album_peak: f64,
-}
-
-#[derive(Default)]
-struct Options {
-    // Gain options
-    gain_steps: Option<i32>,              // -g <i>
-    gain_modifier_db: f64,                // -d <n>: modify suggested dB gain (mp3gain compatible)
-    channel_gain: Option<(Channel, i32)>, // -l <channel> <gain>
-    gain_modifier: i32,                   // -m <i>: modify suggested gain by integer steps
-
-    // Mode options
-    undo: bool,                     // -u
-    stored_tag_mode: StoredTagMode, // -s <mode>
-    use_id3v2: bool,                // -s i: use ID3v2 tags instead of APEv2
-    track_gain: bool,               // -r (apply track gain)
-    album_gain: bool,               // -a (apply album gain)
-    skip_album: bool,               // -e: skip album analysis
-    max_amplitude_only: bool,       // -x: only find max amplitude
-    track_index: Option<u32>,       // -i <index>: track index for multi-track files
-
-    // Behavior options
-    preserve_timestamp: bool,    // -p
-    ignore_clipping: bool,       // -c
-    prevent_clipping: bool,      // -k
-    quiet: bool,                 // -q
-    recursive: bool,             // -R
-    dry_run: bool,               // -n or --dry-run
-    output_format: OutputFormat, // -o <format>
-    wrap_gain: bool,             // -w: wrap gain values
-    use_temp_file: bool,         // -t: use temp file for writing
-    assume_mpeg2: bool,          // -f: assume MPEG 2 Layer III
-
-    // Files
-    files: Vec<PathBuf>,
-}
-
-impl Options {
-    fn dry_run_prefix(&self) -> &'static str {
-        if self.dry_run {
-            "[DRY RUN] "
-        } else {
-            ""
-        }
-    }
-}
-
-// =============================================================================
-// JSON Output Structures
-// =============================================================================
-
-#[derive(Serialize)]
-struct JsonOutput {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    files: Option<Vec<JsonFileResult>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    album: Option<JsonAlbumResult>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    summary: Option<JsonSummary>,
-}
-
-#[derive(Serialize, Clone, Default)]
-struct JsonFileResult {
-    file: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    status: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    frames: Option<usize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    mpeg_version: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    channel_mode: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    min_gain: Option<u8>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    max_gain: Option<u8>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    avg_gain: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    headroom_steps: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    headroom_db: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    gain_applied_steps: Option<i32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    gain_applied_db: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    loudness_db: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    peak: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    max_amplitude: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    error: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    warning: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    dry_run: Option<bool>,
-}
-
-#[derive(Serialize, Clone, Copy)]
-struct JsonAlbumResult {
-    loudness_db: f64,
-    gain_db: f64,
-    gain_steps: i32,
-    peak: f64,
-}
-
-#[derive(Serialize)]
-struct JsonSummary {
-    total_files: usize,
-    successful: usize,
-    failed: usize,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    dry_run: Option<bool>,
-}
-
-// =============================================================================
-// Main
-// =============================================================================
 
 fn main() -> Result<()> {
     let args: Vec<String> = env::args().collect();
@@ -189,285 +51,6 @@ fn main() -> Result<()> {
 
     let opts = parse_args(&args[1..])?;
     run(opts)
-}
-
-fn parse_args(args: &[String]) -> Result<Options> {
-    let mut opts = Options::default();
-    let mut i = 0;
-
-    while i < args.len() {
-        let arg = &args[i];
-
-        if arg == "--dry-run" {
-            opts.dry_run = true;
-            i += 1;
-            continue;
-        }
-
-        if arg == "--help" {
-            print_usage();
-            std::process::exit(0);
-        }
-
-        if arg == "--version" {
-            print_version();
-            std::process::exit(0);
-        }
-
-        if arg.starts_with('-') && arg.len() > 1 && !arg.starts_with("--") {
-            let flag = &arg[1..];
-
-            match flag {
-                "g" => {
-                    i += 1;
-                    if i >= args.len() {
-                        eprintln!("{}: -g requires an argument", "error".red().bold());
-                        std::process::exit(1);
-                    }
-                    opts.gain_steps = Some(
-                        args[i]
-                            .parse()
-                            .map_err(|_| anyhow::anyhow!("invalid gain value: {}", args[i]))?,
-                    );
-                }
-                "d" => {
-                    // mp3gain compatible: -d modifies the suggested dB gain
-                    // (adjusts target level relative to 89 dB reference)
-                    i += 1;
-                    if i >= args.len() {
-                        eprintln!("{}: -d requires an argument", "error".red().bold());
-                        std::process::exit(1);
-                    }
-                    opts.gain_modifier_db = args[i]
-                        .parse()
-                        .map_err(|_| anyhow::anyhow!("invalid dB value: {}", args[i]))?;
-                }
-                "m" => {
-                    i += 1;
-                    if i >= args.len() {
-                        eprintln!("{}: -m requires an argument", "error".red().bold());
-                        std::process::exit(1);
-                    }
-                    opts.gain_modifier = args[i]
-                        .parse()
-                        .map_err(|_| anyhow::anyhow!("invalid modifier value: {}", args[i]))?;
-                }
-                "s" => {
-                    i += 1;
-                    if i >= args.len() {
-                        eprintln!("{}: -s requires an argument", "error".red().bold());
-                        std::process::exit(1);
-                    }
-                    match args[i].as_str() {
-                        "c" => opts.stored_tag_mode = StoredTagMode::Check,
-                        "d" => opts.stored_tag_mode = StoredTagMode::Delete,
-                        "s" => opts.stored_tag_mode = StoredTagMode::Skip,
-                        "r" => opts.stored_tag_mode = StoredTagMode::Recalc,
-                        "i" => {
-                            opts.use_id3v2 = true;
-                        }
-                        "a" => opts.stored_tag_mode = StoredTagMode::UseApev2,
-                        other => {
-                            eprintln!(
-                                "{}: unknown -s mode '{}', use c/d/s/r/i/a",
-                                "error".red().bold(),
-                                other
-                            );
-                            std::process::exit(1);
-                        }
-                    }
-                }
-                "o" => {
-                    // mp3gain compatibility: -o without argument means TSV output
-                    // Check if next arg is a valid format specifier
-                    let next_is_format = if i + 1 < args.len() {
-                        matches!(
-                            args[i + 1].to_lowercase().as_str(),
-                            "json" | "text" | "tsv" | "db"
-                        )
-                    } else {
-                        false
-                    };
-
-                    if next_is_format {
-                        i += 1;
-                        match args[i].to_lowercase().as_str() {
-                            "json" => opts.output_format = OutputFormat::Json,
-                            "text" => opts.output_format = OutputFormat::Text,
-                            "tsv" | "db" => opts.output_format = OutputFormat::Tsv,
-                            _ => unreachable!(),
-                        }
-                    } else {
-                        // mp3gain compatible: -o alone means TSV
-                        opts.output_format = OutputFormat::Tsv;
-                    }
-                }
-                "l" => {
-                    // -l <channel> <gain> : apply gain to specific channel
-                    i += 1;
-                    if i >= args.len() {
-                        eprintln!(
-                            "{}: -l requires two arguments: <channel> <gain>",
-                            "error".red().bold()
-                        );
-                        std::process::exit(1);
-                    }
-                    let channel_arg: usize = args[i].parse().map_err(|_| {
-                        anyhow::anyhow!(
-                            "invalid channel number: {} (use 0 for left, 1 for right)",
-                            args[i]
-                        )
-                    })?;
-                    let channel = Channel::from_index(channel_arg).ok_or_else(|| {
-                        anyhow::anyhow!(
-                            "invalid channel: {} (use 0 for left, 1 for right)",
-                            channel_arg
-                        )
-                    })?;
-
-                    i += 1;
-                    if i >= args.len() {
-                        eprintln!(
-                            "{}: -l requires two arguments: <channel> <gain>",
-                            "error".red().bold()
-                        );
-                        std::process::exit(1);
-                    }
-                    let gain: i32 = args[i]
-                        .parse()
-                        .map_err(|_| anyhow::anyhow!("invalid gain value: {}", args[i]))?;
-
-                    opts.channel_gain = Some((channel, gain));
-                }
-                "r" => opts.track_gain = true,
-                "a" => opts.album_gain = true,
-                "e" => opts.skip_album = true,
-                "x" => opts.max_amplitude_only = true,
-                "i" => {
-                    i += 1;
-                    if i >= args.len() {
-                        eprintln!("{}: -i requires an argument", "error".red().bold());
-                        std::process::exit(1);
-                    }
-                    opts.track_index = Some(
-                        args[i]
-                            .parse()
-                            .map_err(|_| anyhow::anyhow!("invalid track index: {}", args[i]))?,
-                    );
-                }
-                "u" => opts.undo = true,
-                "p" => opts.preserve_timestamp = true,
-                "c" => opts.ignore_clipping = true,
-                "k" => opts.prevent_clipping = true,
-                "q" => opts.quiet = true,
-                "R" => opts.recursive = true,
-                "n" => opts.dry_run = true,
-                "w" => opts.wrap_gain = true,
-                "t" => opts.use_temp_file = true,
-                "f" => opts.assume_mpeg2 = true,
-                "v" | "-version" => {
-                    print_version();
-                    std::process::exit(0);
-                }
-                "h" | "-help" => {
-                    print_usage();
-                    std::process::exit(0);
-                }
-                // Handle combined short flags like -qp, -kc, etc.
-                _ if flag.chars().all(|c| "pqckuranRewxtf".contains(c)) => {
-                    for c in flag.chars() {
-                        match c {
-                            'p' => opts.preserve_timestamp = true,
-                            'q' => opts.quiet = true,
-                            'c' => opts.ignore_clipping = true,
-                            'k' => opts.prevent_clipping = true,
-                            'u' => opts.undo = true,
-                            'r' => opts.track_gain = true,
-                            'a' => opts.album_gain = true,
-                            'n' => opts.dry_run = true,
-                            'R' => opts.recursive = true,
-                            'e' => opts.skip_album = true,
-                            'w' => opts.wrap_gain = true,
-                            'x' => opts.max_amplitude_only = true,
-                            't' => opts.use_temp_file = true,
-                            'f' => opts.assume_mpeg2 = true,
-                            _ => {}
-                        }
-                    }
-                }
-                // Handle -g with attached value (e.g., -g2)
-                _ if flag.starts_with('g') => {
-                    let val = &flag[1..];
-                    opts.gain_steps = Some(
-                        val.parse()
-                            .map_err(|_| anyhow::anyhow!("invalid gain value: {}", val))?,
-                    );
-                }
-                // Handle -d with attached value (e.g., -d4.5)
-                _ if flag.starts_with('d') => {
-                    let val = &flag[1..];
-                    opts.gain_modifier_db = val
-                        .parse()
-                        .map_err(|_| anyhow::anyhow!("invalid dB value: {}", val))?;
-                }
-                // Handle -m with attached value (e.g., -m2)
-                _ if flag.starts_with('m') => {
-                    let val = &flag[1..];
-                    opts.gain_modifier = val
-                        .parse()
-                        .map_err(|_| anyhow::anyhow!("invalid modifier value: {}", val))?;
-                }
-                // Handle -i with attached value (e.g., -i1)
-                _ if flag.starts_with('i') => {
-                    let val = &flag[1..];
-                    opts.track_index = Some(
-                        val.parse()
-                            .map_err(|_| anyhow::anyhow!("invalid track index: {}", val))?,
-                    );
-                }
-                _ => {
-                    eprintln!("{}: unknown option: -{}", "warning".yellow().bold(), flag);
-                }
-            }
-        } else if !arg.starts_with("--") {
-            // It's a file
-            opts.files.push(PathBuf::from(arg));
-        }
-
-        i += 1;
-    }
-
-    Ok(opts)
-}
-
-fn expand_files_recursive(paths: &[PathBuf]) -> Result<Vec<PathBuf>> {
-    let mut result = Vec::new();
-
-    for path in paths {
-        if path.is_dir() {
-            collect_audio_files(path, &mut result)?;
-        } else {
-            result.push(path.clone());
-        }
-    }
-
-    result.sort();
-    Ok(result)
-}
-
-fn collect_audio_files(dir: &Path, result: &mut Vec<PathBuf>) -> Result<()> {
-    for entry in std::fs::read_dir(dir)? {
-        let entry = entry?;
-        let path = entry.path();
-
-        if path.is_dir() {
-            collect_audio_files(&path, result)?;
-        } else if mp3rgain::is_supported_audio_path(&path) {
-            result.push(path);
-        }
-    }
-
-    Ok(())
 }
 
 fn run(mut opts: Options) -> Result<()> {
@@ -537,81 +120,6 @@ fn run(mut opts: Options) -> Result<()> {
         // Default: analyze files (mp3gain compatible)
         // With -d modifier, perform ReplayGain analysis
         cmd_info(&opts.files, &opts)
-    }
-}
-
-// =============================================================================
-// Progress Bar
-// =============================================================================
-
-fn create_progress_bar(total: usize, opts: &Options) -> Option<ProgressBar> {
-    if opts.quiet || opts.output_format != OutputFormat::Text || total < PROGRESS_THRESHOLD {
-        return None;
-    }
-
-    let pb = ProgressBar::new(total as u64);
-    pb.set_style(
-        ProgressStyle::default_bar()
-            .template("{spinner:.cyan} [{bar:40.cyan/blue}] {pos}/{len} {msg}")
-            .unwrap()
-            .progress_chars("=>-"),
-    );
-    Some(pb)
-}
-
-fn progress_set_message(pb: &Option<ProgressBar>, msg: &str) {
-    if let Some(ref pb) = pb {
-        pb.set_message(msg.to_string());
-    }
-}
-
-fn progress_inc(pb: &Option<ProgressBar>) {
-    if let Some(ref pb) = pb {
-        pb.inc(1);
-    }
-}
-
-fn progress_finish(pb: Option<ProgressBar>) {
-    if let Some(pb) = pb {
-        pb.finish_and_clear();
-    }
-}
-
-/// Minimum file size (1 MB) to show per-file analysis progress bar
-const ANALYSIS_PROGRESS_MIN_SIZE: u64 = 1_000_000;
-
-fn create_analysis_progress_bar(
-    mp: &MultiProgress,
-    file: &Path,
-    opts: &Options,
-) -> Option<ProgressBar> {
-    if opts.quiet || opts.output_format != OutputFormat::Text {
-        return None;
-    }
-    let file_size = std::fs::metadata(file).map(|m| m.len()).unwrap_or(0);
-    if file_size < ANALYSIS_PROGRESS_MIN_SIZE {
-        return None;
-    }
-    let pb = mp.add(ProgressBar::new(file_size));
-    pb.set_style(
-        ProgressStyle::default_bar()
-            .template("      [{bar:30.cyan/blue}] {bytes}/{total_bytes}")
-            .unwrap()
-            .progress_chars("=>-"),
-    );
-    Some(pb)
-}
-
-fn update_analysis_progress(pb: &Option<ProgressBar>, bytes_read: u64, total_bytes: u64) {
-    if let Some(ref pb) = pb {
-        pb.set_length(total_bytes);
-        pb.set_position(bytes_read);
-    }
-}
-
-fn finish_analysis_progress(pb: Option<ProgressBar>) {
-    if let Some(pb) = pb {
-        pb.finish_and_clear();
     }
 }
 
@@ -2796,103 +2304,4 @@ fn restore_timestamp(file: &Path, mtime: SystemTime) {
         .write(true)
         .open(file)
         .and_then(|f| f.set_times(std::fs::FileTimes::new().set_modified(mtime)));
-}
-
-// =============================================================================
-// Help / Version
-// =============================================================================
-
-fn print_version() {
-    println!("mp3rgain version {}", VERSION);
-    println!("A modern mp3gain replacement written in Rust");
-    println!();
-    println!("Each gain step = {} dB", GAIN_STEP_DB);
-}
-
-fn print_usage() {
-    println!("{} version {}", "mp3rgain".green().bold(), VERSION);
-    println!("Lossless MP3 volume adjustment - a modern mp3gain replacement");
-    println!();
-    println!("{}", "USAGE:".cyan().bold());
-    println!("    mp3rgain [OPTIONS] <FILES>...");
-    println!();
-    println!("{}", "OPTIONS:".cyan().bold());
-    println!(
-        "    -g <i>      Apply gain of i steps (each step = {} dB)",
-        GAIN_STEP_DB
-    );
-    println!("    -d <n>      Apply gain of n dB (rounded to nearest step)");
-    println!("    -l <c> <g>  Apply gain to left (0) or right (1) channel only");
-    println!("    -m <i>      Modify suggested gain by integer i");
-    println!("    -r          Apply Track gain (ReplayGain analysis)");
-    println!("    -a          Apply Album gain (ReplayGain analysis)");
-    println!("    -e          Skip album analysis (even with multiple files)");
-    println!("    -i <n>      Specify which audio track to process (default: 0)");
-    println!("    -u          Undo gain changes (restore from APEv2 tag)");
-    println!("    -x          Only find max amplitude of file");
-    println!("    -s <mode>   Stored tag handling:");
-    println!("                  c = check/show stored tag info");
-    println!("                  d = delete stored tag info");
-    println!("                  s = skip (ignore) stored tag info");
-    println!("                  r = force recalculation");
-    println!("                  i = use ID3v2 tags (not fully supported)");
-    println!("                  a = use APEv2 tags (default)");
-    println!("    -p          Preserve original file timestamp");
-    println!("    -c          Ignore clipping warnings");
-    println!("    -k          Prevent clipping (automatically limit gain)");
-    println!("    -w          Wrap gain values (instead of clamping)");
-    println!("    -t          Use temp file for writing (safer, required for some ops)");
-    println!("    -f          Assume MPEG 2 Layer III (compatibility, no effect)");
-    println!("    -q          Quiet mode (less output)");
-    println!("    -R          Process directories recursively");
-    println!("    -n          Dry-run mode (show what would be done)");
-    println!("    --dry-run   Same as -n");
-    println!("    -o <fmt>    Output format: 'text' (default), 'json', or 'tsv'");
-    println!("    -v          Show version");
-    println!("    -h          Show this help");
-    println!();
-    println!("{}", "EXAMPLES:".cyan().bold());
-    println!("    mp3rgain song.mp3              Show file info");
-    println!("    mp3rgain -g 2 song.mp3         Apply +2 steps (+3.0 dB)");
-    println!("    mp3rgain -g -3 song.mp3        Apply -3 steps (-4.5 dB)");
-    println!("    mp3rgain -d 4.5 song.mp3       Apply +4.5 dB (rounds to +3 steps)");
-    println!("    mp3rgain -r song.mp3           Analyze and apply track gain");
-    println!("    mp3rgain -a *.mp3              Analyze and apply album gain");
-    println!("    mp3rgain -r -m 2 *.mp3         Apply track gain + 2 steps");
-    println!("    mp3rgain -e *.mp3              Track gain only (skip album calc)");
-    println!("    mp3rgain -u song.mp3           Undo previous gain changes");
-    println!("    mp3rgain -x song.mp3           Show max amplitude only");
-    println!("    mp3rgain -s c *.mp3            Check stored tag info");
-    println!("    mp3rgain -s d *.mp3            Delete stored tag info");
-    println!("    mp3rgain -g 2 -p song.mp3      Apply gain, preserve timestamp");
-    println!("    mp3rgain -k -g 5 song.mp3      Apply gain with clipping prevention");
-    println!("    mp3rgain -w -g 10 song.mp3     Apply gain with wrapping");
-    println!("    mp3rgain -t -g 2 song.mp3      Apply gain using temp file");
-    println!("    mp3rgain -R /path/to/music     Process directory recursively");
-    println!("    mp3rgain -n -g 2 *.mp3         Dry-run (preview changes)");
-    println!("    mp3rgain -o json song.mp3      Output in JSON format");
-    println!("    mp3rgain -o tsv *.mp3          Output in tab-separated format");
-    println!("    mp3rgain -l 0 3 song.mp3       Apply +3 steps to left channel");
-    println!("    mp3rgain -l 1 -2 song.mp3      Apply -2 steps to right channel");
-    println!();
-    println!("{}", "NOTES:".cyan().bold());
-    println!(
-        "    - Each gain step = {} dB (fixed by MP3 specification)",
-        GAIN_STEP_DB
-    );
-    println!("    - Changes are lossless and reversible");
-    println!("    - Gain changes are stored in APEv2 tags for undo support");
-    println!("    - Progress bar shown automatically for 5+ files");
-    if replaygain::is_available() {
-        println!(
-            "    - ReplayGain analysis is {} (target: {} dB)",
-            "enabled".green(),
-            REPLAYGAIN_REFERENCE_DB
-        );
-    } else {
-        println!();
-        println!("{}", "REPLAYGAIN:".yellow().bold());
-        println!("    -r and -a options require the 'replaygain' feature:");
-        println!("    cargo install mp3rgain --features replaygain");
-    }
 }
